refactor(daemon): delete external-skills-bootstrap.ts; lazy external path is sole entry
Part of plan: skill-isolation.md (PR 33 of 34)
