+++ AGENTS.md
@@ -166,9 +166,9 @@ Each LLM call site has a stable identifier (`LLMCallSite` from `assistant/src/co
 
 ## Skill Isolation
 
-The `assistant/` module must not import from `skills/` via relative paths (e.g. `../skills/meet-join/...`). The Docker build context only exposes a curated subset of `skills/` — whitelisted by the repo-root `.dockerignore` — so an import reaching into an unlisted skill path breaks at runtime. Skills wire into the assistant through registries (`registerShutdownHook`, `registerSkillRoute`, `registerExternalTools`) — the assistant provides the hooks, the skill calls them. See `skills/meet-join/AGENTS.md` for the meet-join isolation rules.
+The `assistant/` module must not import from `skills/` via relative paths (e.g. `../skills/meet-join/...`), and `skills/` must not import from `assistant/`. Both directions are enforced by `assistant/src/__tests__/skill-boundary-guard.test.ts`.
 
-There is one narrow exception: `assistant/src/daemon/external-skills-bootstrap.ts` is the single file in `assistant/` permitted to import from `skills/`, and only as named `register(host)` imports of a first-party skill's `register.ts`. The exception exists because `bun --compile` only traces statically analyzed imports into the final binary (dynamic relative imports fail inside `/$bunfs/`), so a skill whose tools must be visible to the LLM at daemon startup has to be reachable from a static import somewhere in `assistant/`. The bootstrap builds a `DaemonSkillHost` (concrete implementation of the `SkillHost` contract in `@vellumai/skill-host-contracts`) and passes it to each skill's `register(host)`; skills use the host to access loggers, config, registries, and providers without importing anything from `assistant/`. Any skill referenced from this file must (a) be first-party, (b) have its runtime files whitelisted in the repo-root `.dockerignore` so the assistant Docker build context includes them (the `Dockerfile` itself copies `skills/` generically and does not name individual skills), and (c) export nothing from `register.ts` other than the `register(host: SkillHost): void` entry point consumed by the bootstrap. All other `assistant/` → `skills/` imports remain forbidden.
+First-party skills run as separate processes. The daemon ships their source tree alongside its binary (Docker build context whitelisted by the repo-root `.dockerignore`; macOS `.app` Resources/) plus a generated `manifest.json` that lists the skill's tools, routes, and shutdown hooks. At daemon startup, `assistant/src/daemon/meet-host-startup.ts` reads the manifest and installs proxy tools/routes/hooks; on first invocation, `MeetHostSupervisor` spawns the skill via `bun run` and dispatches via the `assistant-skill.sock` IPC socket. The skill speaks to the daemon through the `SkillHost` contract in `@vellumai/skill-host-contracts` — neither side imports the other. See `skills/meet-join/AGENTS.md` for the meet-join-specific shape.
 
 ## Tooling Direction
